requirements.txt
beautifulsoup4==4.12.3
bibtexparser==1.4.2
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crossrefapi==1.6.0
doi2bib==0.4.0
Flask==3.1.0
Flask-WTF==1.2.2
future==1.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pyparsing==3.2.0
requests==2.32.3
soupsieve==2.6
urllib3==1.26.16
Werkzeug==3.1.3
WTForms==3.2.1
